+++ common/src/main/java/cf/baradist/AbstractPathFinder.java
@@ -4,14 +4,20 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Arrays;
+import java.util.Set;
+import java.util.Stack;
 import java.util.stream.Collectors;
 
 import static java.lang.Double.NaN;
 
-public class AbstractPathFinder {
+public abstract class AbstractPathFinder {
     protected Double[][] matrix;
+    protected Set<Integer> visitedIndexes;
+    protected Double[] d;
     protected int size;
     protected int from;
+
+    abstract public void findWeights(int from, int to);
 
     public void readGraphMatrix(String filename) {
         try {
@@ -24,6 +30,29 @@ public class AbstractPathFinder {
         } catch (IOException e) {
             throw new RuntimeException(e);
         }
+    }
+
+    public Stack<Edge> findPath(int to) {
+        int current = to;
+        int currentFrom;
+        Stack<Edge> path = new Stack<>();
+        while (current != from) {
+            currentFrom = findPrev(current);
+            path.push(new Edge(currentFrom, current, matrix[currentFrom][current]));
+            current = currentFrom;
+        }
+        return path;
+    }
+
+    private int findPrev(int current) {
+        for (int i = 0; i < size; i++) {
+            if (isRealValue(matrix[i][current]) && visitedIndexes.contains(i)) {
+                if (d[current] == d[i] + matrix[i][current]) {
+                    return i;
+                }
+            }
+        }
+        throw new IllegalArgumentException("Can't find a path");
     }
 
     public int getSize() {
@@ -54,7 +83,7 @@ public class AbstractPathFinder {
                 .collect(Collectors.joining("\t"));
     }
 
-    protected static String toString(Double value) {
+    private static String toString(Double value) {
         return value == null ? "null" : value.isInfinite() ? "∞" : String.valueOf(value);
     }
 }
